create new Emoji component and leave the old implementation as OldEmoji
