+++ src/components/Callout.tsx
@@ -5,7 +5,7 @@ import { GatsbyImage } from "gatsby-plugin-image"
 import Translation from "./Translation"
 import { TranslationKey } from "../utils/translations"
 // Components
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 
 const StyledCard = styled.div`
   display: flex;
