+++ src/components/Nav/Mobile.tsx
@@ -3,7 +3,7 @@ import styled from "@emotion/styled"
 import { useIntl } from "react-intl"
 import { motion } from "framer-motion"
 
-import Emoji from "../Emoji"
+import Emoji from "../OldEmoji"
 import Icon from "../Icon"
 import Link from "../Link"
 import NakedButton from "../NakedButton"
