+++ src/components/SectionNav.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import styled from "@emotion/styled"
 
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import Translation from "./Translation"
 
 const Card = styled.div`
