+++ src/components/StablecoinAccordion.tsx
@@ -7,7 +7,7 @@ import styled from "@emotion/styled"
 // import { motion } from "framer-motion"
 import ButtonLink from "./ButtonLink"
 import CardList, { CardListItem } from "./CardList"
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import Link, { navigate } from "./Link"
 import Icon from "./Icon"
 import InfoBanner from "./InfoBanner"
