+++ src/components/TranslationBanner.tsx
@@ -3,7 +3,7 @@ import styled from "@emotion/styled"
 
 import ButtonLink from "./ButtonLink"
 import Icon from "./Icon"
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import Translation from "./Translation"
 
 const H3 = styled.h3`
