+++ src/components/TranslationBannerLegal.tsx
@@ -4,7 +4,7 @@ import styled from "@emotion/styled"
 
 // Component
 import { ButtonPrimary } from "./SharedStyledComponents"
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import Icon from "./Icon"
 import Translation from "./Translation"
