+++ src/components/TranslationLeaderboard.tsx
@@ -4,7 +4,7 @@ import styled from "@emotion/styled"
 import { reverse, sortBy } from "lodash"
 
 // Components
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import { Option, OptionContainer, OptionText } from "./SharedStyledComponents"
 import Translation from "./Translation"
