+++ src/components/UpgradeBannerNotification.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import BannerNotification from "./BannerNotification"
-import Emoji from "./Emoji"
+import Emoji from "./OldEmoji"
 import Link from "./Link"
 import styled from "@emotion/styled"
