+++ src/pages/stablecoins.tsx
@@ -7,7 +7,7 @@ import { useIntl } from "react-intl"
 import ButtonLink from "../components/ButtonLink"
 import CalloutBanner from "../components/CalloutBanner"
 import DataProductCard from "../components/DataProductCard"
-import Emoji from "../components/Emoji"
+import Emoji from "../components/OldEmoji"
 import FeedbackCard from "../components/FeedbackCard"
 import GhostCard from "../components/GhostCard"
 import HorizontalCard from "../components/HorizontalCard"
