+++ src/pages/staking/deposit-contract.tsx
@@ -10,7 +10,7 @@ import ButtonLink from "../../components/ButtonLink"
 import CardList from "../../components/CardList"
 import Checkbox from "../../components/Checkbox"
 import CopyToClipboard from "../../components/CopyToClipboard"
-import Emoji from "../../components/Emoji"
+import Emoji from "../../components/OldEmoji"
 import InfoBanner from "../../components/InfoBanner"
 import Link from "../../components/Link"
 import PageMetadata from "../../components/PageMetadata"
