+++ src/pages/upgrades/get-involved/index.tsx
@@ -9,7 +9,7 @@ import { translateMessageId, TranslationKey } from "../../../utils/translations"
 import Card from "../../../components/Card"
 import Leaderboard, { Person } from "../../../components/Leaderboard"
 import CalloutBanner from "../../../components/CalloutBanner"
-import Emoji from "../../../components/Emoji"
+import Emoji from "../../../components/OldEmoji"
 import ProductCard from "../../../components/ProductCard"
 import ButtonLink from "../../../components/ButtonLink"
 import PageMetadata from "../../../components/PageMetadata"
