+++ src/pages/upgrades/index.tsx
@@ -8,7 +8,7 @@ import ButtonLink from "../../components/ButtonLink"
 import Card from "../../components/Card"
 import ActionCard from "../../components/ActionCard"
 import CalloutBanner from "../../components/CalloutBanner"
-import Emoji from "../../components/Emoji"
+import Emoji from "../../components/OldEmoji"
 import UpgradeArticles from "../../components/UpgradeArticles"
 import ExpandableCard from "../../components/ExpandableCard"
 import GhostCard from "../../components/GhostCard"
